Renamed /editor_input_delegation to /editor_abstractions to recognize the growing scope of this spike.
