+++ example/lib/spikes/editor_abstractions/composition/document_composer_actions.dart
@@ -1,9 +1,9 @@
-import 'package:example/spikes/editor_input_delegation/document/document_editor.dart';
-import 'package:example/spikes/editor_input_delegation/document/document_nodes.dart';
-import 'package:example/spikes/editor_input_delegation/document/rich_text_document.dart';
-import 'package:example/spikes/editor_input_delegation/selectable_text/selectable_text.dart';
-import 'package:example/spikes/editor_input_delegation/layout/document_layout.dart';
-import 'package:example/spikes/editor_input_delegation/selection/editor_selection.dart';
+import 'package:example/spikes/editor_abstractions/document/document_editor.dart';
+import 'package:example/spikes/editor_abstractions/document/document_nodes.dart';
+import 'package:example/spikes/editor_abstractions/document/rich_text_document.dart';
+import 'package:example/spikes/editor_abstractions/selectable_text/selectable_text.dart';
+import 'package:example/spikes/editor_abstractions/layout/document_layout.dart';
+import 'package:example/spikes/editor_abstractions/selection/editor_selection.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter/widgets.dart';
