+++ example/lib/spikes/editor_abstractions/document/document_editor.dart
