+++ example/lib/spikes/editor_abstractions/document/document_nodes.dart
