+++ example/lib/spikes/editor_abstractions/document/rich_text_document.dart
