+++ example/lib/spikes/editor_abstractions/ui_components/image.dart
